fix(Dashboard): fix underflow

+++ contracts/0.8.25/vaults/dashboard/NodeOperatorFee.sol
@@ -147,7 +147,7 @@ contract NodeOperatorFee is Permissions {
         // cast down safely clamping to int128.max
         int128 adjustment = int128(int256(accruedRewardsAdjustment & ADJUSTMENT_CLAMP_MASK));
 
-        int128 rewardsAccrued = int128(latestReport.valuation - _lastClaimedReport.valuation) -
+        int128 rewardsAccrued = int128(latestReport.valuation) - int128(_lastClaimedReport.valuation) -
             (latestReport.inOutDelta - _lastClaimedReport.inOutDelta) -
             adjustment;
